feat: edit the structure of the app
- Separate the logic from the display
- Add the structure for nbgl display
- Add some github workflows
