+++ .github/workflows/ci.yml
@@ -1,53 +1,53 @@
-name: CI
+# name: CI
 
-on:
-  # Triggers the workflow on push or pull request events but only for the master branch
-  push:
-    branches: [ master, develop ]
-  pull_request:
-    branches: [ master, develop ]
-  # Allows you to run this workflow manually from the Actions tab
-  workflow_dispatch:
+# on:
+#   # Triggers the workflow on push or pull request events but only for the master branch
+#   push:
+#     branches: [ master, develop ]
+#   pull_request:
+#     branches: [ master, develop ]
+#   # Allows you to run this workflow manually from the Actions tab
+#   workflow_dispatch:
 
-jobs:
-  nano-build:
-    runs-on: ubuntu-latest
-    container:
-      image: ghcr.io/ledgerhq/ledger-app-builder/ledger-app-builder:latest
-    strategy:
-      matrix:
-        sdk:
-        - path: $NANOS_SDK
-          name: nanos
-        - path: $NANOX_SDK
-          name: nanox
-        - path: $NANOSP_SDK
-          name: nanosp
-    steps:
-      - uses: actions/checkout@v2
-      - name: Build application for SDK ${{ matrix.sdk.name }}
-        run: |
-          make clean
-          make BOLOS_SDK=${{ matrix.sdk.path }}
-      - name: Upload app binary
-        uses: actions/upload-artifact@v2
-        with:
-          name: app
-          path: bin
+# jobs:
+#   nano-build:
+#     runs-on: ubuntu-latest
+#     container:
+#       image: ghcr.io/ledgerhq/ledger-app-builder/ledger-app-builder:latest
+#     strategy:
+#       matrix:
+#         sdk:
+#         - path: $NANOS_SDK
+#           name: nanos
+#         - path: $NANOX_SDK
+#           name: nanox
+#         - path: $NANOSP_SDK
+#           name: nanosp
+#     steps:
+#       - uses: actions/checkout@v2
+#       - name: Build application for SDK ${{ matrix.sdk.name }}
+#         run: |
+#           make clean
+#           make BOLOS_SDK=${{ matrix.sdk.path }}
+#       - name: Upload app binary
+#         uses: actions/upload-artifact@v2
+#         with:
+#           name: app
+#           path: bin
 
-  scan-build:
-    name: Clang Static Analyzer
-    runs-on: ubuntu-latest
-    container:
-      image: ghcr.io/ledgerhq/ledger-app-builder/ledger-app-builder:latest
-    steps:
-      - uses: actions/checkout@v2
-      - name: Build with Clang Static Analyzer
-        run: |
-          make clean
-          scan-build --use-cc=clang -analyze-headers -enable-checker security -enable-checker unix -enable-checker valist -o scan-build --status-bugs make default
-      - uses: actions/upload-artifact@v2
-        if: failure()
-        with:
-          name: scan-build
-          path: scan-build
+#   scan-build:
+#     name: Clang Static Analyzer
+#     runs-on: ubuntu-latest
+#     container:
+#       image: ghcr.io/ledgerhq/ledger-app-builder/ledger-app-builder:latest
+#     steps:
+#       - uses: actions/checkout@v2
+#       - name: Build with Clang Static Analyzer
+#         run: |
+#           make clean
+#           scan-build --use-cc=clang -analyze-headers -enable-checker security -enable-checker unix -enable-checker valist -o scan-build --status-bugs make default
+#       - uses: actions/upload-artifact@v2
+#         if: failure()
+#         with:
+#           name: scan-build
+#           path: scan-build
