+++ Makefile
@@ -29,8 +29,8 @@ endif
 APP_LOAD_PARAMS += $(COMMON_LOAD_PARAMS)
 
 # Pending review parameters
-APP_LOAD_PARAMS += --tlvraw 9F:01
-DEFINES += HAVE_PENDING_REVIEW_SCREEN
+# APP_LOAD_PARAMS += --tlvraw 9F:01
+# DEFINES += HAVE_PENDING_REVIEW_SCREEN
 
 ##################
 # Define Version #
