+++ ledger_app.toml
@@ -1,4 +1,4 @@
 [app]
 build_directory = "./"
 sdk = "C"
-devices = ["nanos", "nanox", "nanos+"]
+devices = ["nanos", "nanos+", "nanox", "flex", "stax"]
