+++ src/handler/get_app_configuration.c
