+++ src/main.c
@@ -288,7 +288,7 @@ __attribute__((section(".boot"))) int main(void) {
                 USB_power(0);
                 USB_power(1);
 
-                ui_idle();
+                ui_main_menu();
 
 #ifdef HAVE_BLE
                 BLE_power(0, NULL);
