--- src/menu.h
@@ -1,9 +1,0 @@
-#include "globals.h"
-#include "glyphs.h"
-
-#ifndef _MENU_H_
-#define _MENU_H_
-
-void ui_idle(void);
-
-#endif
