+++ src/ui/display.h
@@ -1,0 +1,7 @@
+#pragma once
+
+void ui_display_address();
+void ui_display_public_key();
+void ui_display_sign_transaction(int flow);
+void ui_display_sign();
+
