+++ src/ui/menu.h
@@ -1,0 +1,8 @@
+#include "globals.h"
+#include "glyphs.h"
+
+#pragma once
+
+void ui_main_menu(void);
+
+void ui_about_menu(void);
