+++ src/ui/menu_bagl.c
@@ -1,3 +1,5 @@
+#ifdef HAVE_BAGL
+
 #include "menu.h"
 #include "os.h"
 
@@ -31,10 +33,12 @@ UX_FLOW(ux_idle_flow,
   FLOW_LOOP
 );
 
-void ui_idle(void) {
+void ui_main_menu(void) {
     // reserve a display stack slot if none yet
     if(G_ux.stack_count == 0) {
         ux_stack_push();
     }
     ux_flow_init(0, ux_idle_flow, NULL);
 }
+
+#endif
