+++ src/ui/menu_nbgl.c
@@ -1,0 +1,6 @@
+#ifdef HAVE_NBGL
+
+#include "menu.h"
+
+
+#endif
