+++ src/utils.c
@@ -75,7 +75,7 @@ void send_response(uint8_t tx, bool approve) {
     // Send back the response, do not restart the event loop
     io_exchange(CHANNEL_APDU | IO_RETURN_AFTER_TX, tx);
     // Display back the original UX
-    ui_idle();
+    ui_main_menu();
 }
 
 unsigned int ui_prepro(const bagl_element_t *element) {
@@ -210,3 +210,72 @@ uint8_t convert_hex_amount_to_displayable(const uint8_t* amount, uint8_t decimal
     }
     return targetOffset;
 }
+
+uint8_t set_result_get_address() {
+    uint8_t tx = 0;
+    G_io_apdu_buffer[tx++] = ADDRESS_LENGTH;
+    memmove(G_io_apdu_buffer + tx, data_context.addr_context.address, ADDRESS_LENGTH);
+    tx += ADDRESS_LENGTH;
+    reset_app_context();
+    return tx;
+}
+
+uint8_t set_result_get_public_key() {
+    uint8_t tx = 0;
+    G_io_apdu_buffer[tx++] = PUBLIC_KEY_LENGTH;
+    memmove(G_io_apdu_buffer + tx, data_context.pk_context.public_key, PUBLIC_KEY_LENGTH);
+    tx += PUBLIC_KEY_LENGTH;
+    reset_app_context();
+    return tx;
+}
+
+uint8_t set_result_sign_transaction() {
+    cx_ecfp_private_key_t privateKey;
+    SignTransactionContext_t* context = &data_context.sign_tr_context;
+
+    BEGIN_TRY {
+        TRY {
+            get_private_key(context->account_number, &privateKey);
+            if (!context->sign_with_chain_id) {
+                cx_eddsa_sign(&privateKey, CX_LAST, CX_SHA512, context->to_sign, TO_SIGN_LENGTH, NULL, 0, context->signature, SIGNATURE_LENGTH, NULL);
+            } else {
+                cx_eddsa_sign(&privateKey, CX_LAST, CX_SHA512, context->to_sign, CHAIN_ID_LENGTH + TO_SIGN_LENGTH, NULL, 0, context->signature, SIGNATURE_LENGTH, NULL);
+            }
+        } FINALLY {
+            memset(&privateKey, 0, sizeof(privateKey));
+        }
+    }
+    END_TRY;
+
+    uint8_t tx = 0;
+    G_io_apdu_buffer[tx++] = SIGNATURE_LENGTH;
+    memmove(G_io_apdu_buffer + tx, context->signature, SIGNATURE_LENGTH);
+    tx += SIGNATURE_LENGTH;
+    return tx;
+}
+
+uint8_t set_result_sign() {
+    cx_ecfp_private_key_t privateKey;
+    SignContext_t* context = &data_context.sign_context;
+
+    BEGIN_TRY {
+        TRY {
+            get_private_key(context->account_number, &privateKey);
+            if (!context->sign_with_chain_id) {
+                cx_eddsa_sign(&privateKey, CX_LAST, CX_SHA512, context->to_sign, TO_SIGN_LENGTH, NULL, 0, context->signature, SIGNATURE_LENGTH, NULL);
+            } else {
+                cx_eddsa_sign(&privateKey, CX_LAST, CX_SHA512, context->to_sign, CHAIN_ID_LENGTH + TO_SIGN_LENGTH, NULL, 0, context->signature, SIGNATURE_LENGTH, NULL);
+            }
+        } FINALLY {
+            memset(&privateKey, 0, sizeof(privateKey));
+        }
+    }
+    END_TRY;
+
+    uint8_t tx = 0;
+    G_io_apdu_buffer[tx++] = SIGNATURE_LENGTH;
+    memmove(G_io_apdu_buffer + tx, context->signature, SIGNATURE_LENGTH);
+    tx += SIGNATURE_LENGTH;
+    return tx;
+}
+
